requirements.txt
notebook
pandas
torch
torchvision
torchaudio
python-dotenv
requests
openai
tiktoken
tqdm
Flask
chromadb
python-frontmatter
unstructured[local-inference]
gunicorn
langchain
llama-index
google-search-results
wolframalpha
wikipedia
discord.py
pyowm
OpenWeatherMap
markdown
markdown[extra]
colored_traceback
shell-gpt

numpy
torch>=2.0.1
safetensors==0.3.1
sentencepiece>=0.1.97
ninja==1.11.1
llama_hub